refactor: MovieAPIData 위치 변경 및 Movie type key값 변경
