+++ src/api/get.ts
@@ -1,5 +1,22 @@
 const BASE_URL = 'https://api.themoviedb.org/3';
 
+export interface MovieAPIData {
+  adult: boolean;
+  backdrop_path: string;
+  genre_ids: number[];
+  id: number;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  popularity: number;
+  poster_path: string;
+  release_date: string;
+  title: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
+}
+
 export const popularMovieDataFetchFuncGenerator = () => {
   let currentPage = 1;
