+++ src/type/Movie.ts
@@ -1,23 +1,6 @@
 export interface Movie {
   id: number;
   title: string;
-  poster_path: string;
-  vote_average: number;
-}
-
-export interface MovieAPIData {
-  adult: boolean;
-  backdrop_path: string;
-  genre_ids: number[];
-  id: number;
-  original_language: string;
-  original_title: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  release_date: string;
-  title: string;
-  video: boolean;
-  vote_average: number;
-  vote_count: number;
+  posterPath: string;
+  voteAverage: number;
 }
